Refactor policies background jobs.
