+++ internal/codeintel/policies/repomatcher.go
@@ -1,0 +1,45 @@
+package policies
+
+import (
+	"context"
+	"time"
+
+	"github.com/sourcegraph/sourcegraph/internal/conf"
+	"github.com/sourcegraph/sourcegraph/internal/goroutine"
+)
+
+func (s *Service) NewRepoMatcher(interval time.Duration, configurationPolicyMembershipBatchSize int) goroutine.BackgroundRoutine {
+	return goroutine.NewPeriodicGoroutine(context.Background(), interval, goroutine.HandlerFunc(func(ctx context.Context) error {
+		return s.handleRepoMatcherBatch(ctx, configurationPolicyMembershipBatchSize)
+	}))
+}
+
+func (s *Service) handleRepoMatcherBatch(ctx context.Context, batchSize int) error {
+	policies, err := s.SelectPoliciesForRepositoryMembershipUpdate(ctx, batchSize)
+	if err != nil {
+		return err
+	}
+
+	for _, policy := range policies {
+		var patterns []string
+		if policy.RepositoryPatterns != nil {
+			patterns = *policy.RepositoryPatterns
+		}
+
+		var repositoryMatchLimit *int
+		if val := conf.CodeIntelAutoIndexingPolicyRepositoryMatchLimit(); val != -1 {
+			repositoryMatchLimit = &val
+		}
+
+		// Always call this even if patterns are not supplied. Otherwise we run into the
+		// situation where we have deleted all of the patterns associated with a policy
+		// but it still has entries in the lookup table.
+		if err := s.UpdateReposMatchingPatterns(ctx, patterns, policy.ID, repositoryMatchLimit); err != nil {
+			return err
+		}
+
+		s.matcherMetrics.numPoliciesUpdated.Inc()
+	}
+
+	return nil
+}
